(MATH) fixed signed comparaison.

+++ libspider/libspider/common/Math.h
@@ -72,21 +72,21 @@ namespace spider {
             auto neg = (x < 0 && y > 0) || (x > 0 && y < 0);
             auto a = static_cast<uint64_t>((x < 0 ? (-x) : x));
             auto b = static_cast<uint64_t>((y < 0 ? (-y) : y));
-            return static_cast<int64_t>((neg ? -(a / b) : ceilDiv(a, b)));
+            return neg ? -static_cast<int64_t>(a / b) : static_cast<int64_t>(ceilDiv(a, b));
         }
 
         inline std::int64_t floorDiv(std::int64_t x, std::int64_t y) {
             auto neg = (x < 0 && y > 0) || (x > 0 && y < 0);
             auto a = static_cast<uint64_t>((x < 0 ? (-x) : x));
             auto b = static_cast<uint64_t>((y < 0 ? (-y) : y));
-            return static_cast<int64_t>((neg ? -ceilDiv(a, b) : a / b));
+            return neg ? -static_cast<int64_t>(ceilDiv(a, b)) : static_cast<int64_t>(a / b);
         }
 
         inline std::int32_t floorDiv(std::int32_t x, std::int32_t y) {
             auto neg = (x < 0 && y > 0) || (x > 0 && y < 0);
             auto a = static_cast<uint32_t>((x < 0 ? (-x) : x));
             auto b = static_cast<uint32_t>((y < 0 ? (-y) : y));
-            return static_cast<int32_t>((neg ? -ceilDiv(a, b) : a / b));
+            return neg ? -static_cast<int32_t>(ceilDiv(a, b)) : static_cast<int32_t>(a / b);
         }
 
         inline std::int16_t abs(std::int16_t x) {
